Split server parts of 'Rocket' into 'server.rs'.
